--- .context/auto-review-20260221-182219.md
@@ -1,64 +1,0 @@
-# Review Context
-
-## Branch Info
-
-- Base: origin/master
-- Current: draft
-
-## Changed Files Summary
-
-- A src/cli/draft-actions.ts (new file, 113 lines)
-- A src/cli/draft-server.ts (new file, 1372 lines)
-- M src/cli/message-command.ts (added import + subcommand registration)
-- M src/slack/channels.ts (added resolveChannelName function, 40 lines)
-- M README.md (added draft command to command map + examples)
-- M skills/agent-slack/SKILL.md (added draft section)
-- M skills/agent-slack/references/commands.md (added draft command reference)
-- A docs/draft-editor.png (binary screenshot)
-
-## Changed Line Ranges (PR Scope)
-
-| File                                      | Changed Lines     |
-| ----------------------------------------- | ----------------- |
-| src/cli/draft-actions.ts                  | 1-113 (new file)  |
-| src/cli/draft-server.ts                   | 1-1372 (new file) |
-| src/cli/message-command.ts                | 12, 176-205       |
-| src/slack/channels.ts                     | 83-122            |
-| README.md                                 | 64, 176-192       |
-| skills/agent-slack/SKILL.md               | 93-102            |
-| skills/agent-slack/references/commands.md | 47-56             |
-
-## Review Standards Reference
-
-- Follow /review-code standards
-- Focus on: correctness, security, performance, maintainability
-- Priority levels: Critical > High > Medium > Low
-
-## File Categories
-
-### Backend/API (review-code, review-algorithm-architecture)
-
-- src/cli/draft-actions.ts
-- src/cli/draft-server.ts
-- src/cli/message-command.ts
-- src/slack/channels.ts
-
-### Config/Build (review-code)
-
-(none)
-
-### Utility/Common (review-code)
-
-- README.md
-- skills/agent-slack/SKILL.md
-- skills/agent-slack/references/commands.md
-
-## Skipped Issues (Do Not Re-validate)
-
-<!-- Format: [file:line-range] | [severity] | [reason skipped] | [issue summary] -->
-
-## Iteration State
-
-Current iteration: 1
-Last completed phase: Setup
-Files fixed this iteration: []
